Move command related mods to commands
