+++ nym-vpn-core/crates/nym-vpn-account-controller/src/commands/tasks/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+pub(crate) mod register_device;
+pub(crate) mod request_zknym;
+pub(crate) mod sync_account;
+pub(crate) mod sync_device;
