+++ nym-vpn-core/crates/nym-vpn-account-controller/src/lib.rs
@@ -7,9 +7,7 @@
 // 3. request ticketbooks and top up the local credential store
 
 pub mod shared_state;
-pub mod storage_cleanup;
 
-mod command_handler;
 mod commander;
 mod commands;
 mod connectivity;
@@ -19,8 +17,8 @@ mod storage;
 mod ticketbooks;
 
 pub use commander::AccountControllerCommander;
-pub use commands::AccountCommand;
 pub use controller::AccountController;
 pub use error::Error;
 pub use shared_state::{AccountStateSummary, SharedAccountState};
+pub use storage::remove_files_for_account;
 pub use ticketbooks::{AvailableTicketbook, AvailableTicketbooks};
