+++ nym-vpn-core/crates/nym-vpnd/src/command_interface/protobuf/account.rs
@@ -91,11 +91,11 @@ impl From<AccountError> for nym_vpn_proto::AccountError {
                 message: err.to_string(),
                 details: hashmap! {},
             },
-            AccountError::SendCommand { .. } => nym_vpn_proto::AccountError {
-                kind: AccountErrorType::Storage as i32,
-                message: err.to_string(),
-                details: hashmap! {},
-            },
+            //AccountError::SendCommand { .. } => nym_vpn_proto::AccountError {
+            //    kind: AccountErrorType::Storage as i32,
+            //    message: err.to_string(),
+            //    details: hashmap! {},
+            //},
             AccountError::RecvCommand { .. } => nym_vpn_proto::AccountError {
                 kind: AccountErrorType::Storage as i32,
                 message: err.to_string(),
